SVG: Handle company logo controller
The company logo, which is for example shown on interface, uses a special controller
that allows to cache the content and adapt to proper size.

This controller doesn't handle correctly SVG as it uses a bitmap image only method
for guessing mimetype and file extension.

Switching to the same mimetype guessing as in other controller, we handle this properly.