[IMP] mail: add main attachment support on mail.thread
This is done by adding a new field on mail.thread objects and setting it when posting a message with attachments on a model inheriting mail.thread. The main attachment can be changed using the preview widget: when we close the widget, the open attachment becomes the new main one.