[REF] base, web, website: refactors binary_content
This commit refactors binary_content to make it more readable by
splitting the method and to make it more flexible (related
to the refactor of documents odoo/enterprise#3048).

removes several parameters in web/content, web/image and binary_content
following the refactoring of the module documents.

moved the resize function of web/image outside of the method to
make it globally available.

closes: odoo/odoo#28563
task: #1908896