[FIX] purchase: delete picking linked to RFQ
This diff could be used in order to cancel the internal picking
linked to an RFQ. Since it directly cancel the picking that created
the RFQ, you should be care with warehouse not using a multi steps
reception or delivery. Also the behavior for reordering rule and mto
product are the same. So the impact will be the same.