[REF] web: calendar view redesign
Refactor the calendar view, improving design and introducing
several usability improvements, in detail:

- Increase readability using a better colour selection.
- Add a popover containing all the fields specified in the
  <calendar/> element.
- Remove any field from the event card, just print the event’s
  title.

The new calendar popover consist in two sections.
The first one contains the title, date and time. It may display
also the field used to filter/assign colours (if it’s defined as
a <field> in the root <calendar> element).
Date and time are automatically computed according to the user
timing preferences, language and event duration.
Date and time can be hidden adding respectively the ‘hide_date’
and ‘hide_time’ attributes at the root <calendar/> element.

The second section is scrollable and contains all other fields
specified in the root <calendar/> element.