[IMP] mail: improve activity done message template
Feedback should be the first part of the message without tile
to make it more concise.

If the activity has a note, the note should be displayed too
but after the feedback.

We can also remove the field feedback that was not so usefull
since it was only written just before unlink. Feedback is now
passed as another parameter to the template.

Task: 1918392