[IMP] base : (search) decompose saved filter.
Purpose of task is Currently, the user do not know how a saved custom filter has been defined. Unless the go to studio to see the domain (too technical), in doubt, they will often delete the filter to re-do it to be sure of its definition.
Also, there is no way to edit a saved filter.

 - add a fa-search (icon) button that will 'decompose' the saved filter in the searchbar for information/editing.

This commit is related to task ID : 1918482

closes odoo/odoo#