[FIX] test_main_flows: make the tour succeed standalone
The tour was failing when only the module (and its dependencies) was installed
(but was correctly succeeding on runbot).

The issue here was due to an invisible field in the 'Register Payment' wizard.
The field was set invisible with the onchange because there was only one payment
method on the journal.

The module 'account_check_printing' adds another payment method (check) and the
field becomes visible.