[IMP] web_editor,*: Refactoring of the wysiwyg editor and 'html' field
The refactoring allows the update of the library of summernote.
A new structure was created by transforming all the plugins in the library with
the odoo inheritance system. Plugins are easier to implement with the
AbstractPlugin to add odoo behaviors.

The cleanup allowed to remove some of the code that was in web_editor but only
used in website or in mass_mailing. Some part is still in web_editor but will
be moved in a later time.

The refactoring also had to be done in order to create a set of unit tests.
Each behavior can be tested as well as the behaviors performed during keyboard
interactions (for example: Enter, Tab...)

From now on, the methods of the library (in this case it is about summernote)
can no longer be called by the other modules or files. Only the wysiwyg widgets
can access it, to allow easy updates. The wysiwyg object serves as an interface.
and a lot of tests are done. Thus, the behavior of the outside will be assured.

Depending on the options, the snippets will be present or not, the editor will
be in an iframe or not... all this is transparent from the outside. Regarding
iframes, all controllers related to editing have been removed, the new API no
longer needs them. This speeds up the loading speed, facilitates testing and
removes complexity for the same features.

PUBLIC FEATURES

There are several public methods on the Wysiwyg class:
* Wysiwyg.prepare (WidgetParent) which returns a deferred resolved when the
  librarian (xml, lasy, assets...) is loaded.
* Wysiwyg.getRange (DOM) returns the range (selection in the dom)
* Wysiwyg.setRange (startNode, startOffset, endNode, endOffset) which creates a
  range (selection in the dom)
* Wysiwyg.setRangeFromNode (DOM, options) that creates a range from an element
  (option available to select all, start or end)

A jQuery selector is added, it is ":o_editable" which makes it possible to
know if the current element is editable. That is, if it is contained in a
tag with the attribute 'contentEditable = "true"' or in a tag with the class
'o_editable'.
Several methods are also present:
* focusIn: which makes a focus and places the cursor at the beginning of the
  element
* focusInEnd: which makes a focus and places the cursor at the end of the
  element
* selectContent: which makes a focus and selects the content

HTML FIELD

The HTML field can receive different options:
* style-inline: {boolean} which transforms class into inline style when saving
and vice versa during lecture.
* no-attachment: {boolean} prevents the use of attachments (in media dialog)
* cssEdit: {xml_id} to use a template containing the css to loaded in an iframe
when editing
* cssReadonly: {xml_id} to use a template containing the css to loaded into an
iframe when viewing readonly
* snippets: {xml_id} snippets template (can be used with or without cssEdit)
* wrapper: {template} template qweb static (containing a tag id = "wrapper")
that will include the content during editing (removed at save)

MASS MAILING

A widget for mass mailing has been created. There are now two fields 'body_html'
and 'body_arch'.
'body_arch' containing the code with the class without convertion in style,
useful when editing and one with the inline style visible in readonly mode and
is sent by email.
Advantage: no spreading errors, able du update css/theme, able to do more
modification when converting to the inline style so that a maximum of mail
clients have an impeccable rendering.

Co-authored-by: Antoine Guenet (AGE) <age@odoo.com>