[FIX] point_of_sale: correcly handle print error on mobile
I see 3 cases that may happen on some mobile browsers:
- window.print() works and a popup is opened asking for printers.
  e.g.: Google Chrome Mobile
- window.print() is not implemented and we need to catch this exception.
  e.g.: Firefox Mobile
- window.print() is implemented but not supported and nothing happens;
  No exception is triggered.
  e.g.: Jelly, ...

The first two cases were already handled by this commit @4629a4fb0eaced
but also need to show an error for the last case.
We need to check if printing process started after 2 secs.
If not, it's probably because it's not supported.
In this case, window.print() doesn't return any arror or status.

opw-1916984