[IMP] web: add python qweb functionnality to javascript qweb.
Even if the python and javascript qweb should be the same, some differences remains.
This commits add the 'add' and 'remove' (and 'separator') the 'attributes' to allow to add
or remove an element in a list.

Example to add a class:

<attribute name="class" add="wonderfull_class" separator=" "/>

This was possible in pyhton templates, this is now possible with javascript templates.