[IMP] iot_box: make filesystem rw
By adding the right ramdisks it is not nessesery to mount the root as
read only. The ramdisks can be mountend directly by fstab instead of
using a dedicated daemon service.

By making this change many scripts dedicated to the IoT Box can be
simplefied becouse they can use the proper /etc/ directory and are not
forced to remount the root directory for every change that has to be
made to it.