[WIP] base: properly raise an error in uninstallation
This patch contains two solutions for the same problem.

1.
When the uninstallation cause an integrity error, the error should not be
hidden as it may corrupt the database.

e.g. in case of ir.model.fields.selection removal, resetting the value to null
is done for all records that were using this value. If the value is required,
an error must be raised as a manual action is required.
Otherwise, records are using an orphan selection value without any business
logic.

2.
A user-friendly error is raised for ir.model.fields.selection however a
UserError is also used for unlink override we do want to catch (e.g. removing
account.account)