[ADD] hr_pin: user pin and barcode to identify employees
Purpose
=======

Commit f14092c moved the definition of `pin` and `barcode` fields
on `hr.employee` from the module `hr_attendance` to `hr`.
The purpose was to reuse them in ´point_of_sale´
to allow an employee to connect to pos.

But if neither `hr_attendance` nor ´point_of_sale´ is installed,
those fields are present without any pupose which is confusing.

Specification
=============

This commit moves those fields in a separate module which is installed
when `hr_attendance` or `pos_hr` is installed.