[FIX] pos_restaurant: fix table editing broken after jQuery update
The table edition feature of pos_restaurant was using an old lib
that was never updated to newer versions of jQuery by its author.

For unknown reason, this library went unnoticed when we updated our
codebase to jQuery 3.

Since this library is impossible to update per se, I decided to
remove it entirely and replace it by calls to jQuery.ui.

However, the point of this lib was to handle touch events, which
jQuery.ui does not. For this, this commit introduces a new
library which translates touch events to their click counterparts.