[IMP] web, base: put buttons in list group headers
It is now possible to put buttons in the list view group headers. They will
appear next to the header title when the group is opened.

The buttons are specified in the views in a <groupby> tag in the list arch, with
the following structure:
    <groupby name="groupField">
        <button type="object" name="my_method" string="Button1"/>
    </groupby>

When specifying a group on many2one fields, it is possible to add <fields>,
which can be used for modifiers. These fields are thus belonging on the many2one
comodel, like:
    <groupby name="partner_id">
        <field name="name"/> <!-- name of partner_id -->
        <button type="object" name="my_method" string="Button1" attrs="{'invisible': [('name', '=', 'Georges')]}"/>
    </groupby>

These extra fields are thus fetched in batch when grouping on the field.

Part of task 1915702