- changes in models:
account.tax.tag.template and account.tax.report.section have been merged into a common model, defining both the structure of the tax report and the tags to create: account.tax.report line. It essentially is a renaming of account.tax.tag.template with additional fields on top of it.

- + small fixes on l10n_be taxes