[IMP] web: allow mass update in editable list
From an editable list, if you select several lines then change a field, you are
now asked if you want to apply the same change on all selected lines. If you
accept, a write is directly performed (without applying onchange, as the records
are reloaded and the view re-rendered). If you decline, the classic behavior
still occurs.

Part of task 1915702