[FIX] stock: mto > resupply route
Usecase to reproduce:
- Warehouse1 stock product and do not manufacture
- Warehouse2 manufacture and resupply Warehouse1
- Check resupply from WH2, MTO and Manufacture

Create a SO for WH1 -> It will create the moves chain to
Customer -> WH1 -> transit -> WH2 but it will not trigger
the MO. It happens because the routes resupply and mto are
both selected on the prodcut. However it will take resupply
route by default because the default sequence number is lower.
The rule in resupply do not have the propagate parameter set,
that's why the created move do not trigger the following rule
which is the manufacture rule.

Fix it by adding a sequence on resupply route in order to
prioritize MTO route