[IMP] stock: Sequence on warehouse
A lot of object are created with user's company's first warehouse.
For example the default warehouse on a SO is the first warehouse created
in the current company. If the user want to modify the default
warehouse, he has to delete all the warehouse created before the wanted
and recreate them before.

This commit introduce a sequence number and an order on it in order to
modify the default warehouse. it's probably not super userfreindly and
it could be done by different developement depending the case but this
commit introduce a first base in order to allow it without custom code.