[IMP] iotbox: fstab
Modifications on partition table to simplify workflow and guarantee
safety and lifespan of sd card

By making this change the main partition is always writable for the root
user Linux services like apt-get work again after this commit, because of
this future upgrades with new deps can be done without flashing the sd-card. This is
very important to be able to install new IoT Drivers without rewriting the
complete sd card every time.