[IMP] tools,modules,delivery: Add a neutralize switch
This commit add an option when lauching Odoo in command
line. This option "--neuter-mode" is defined by False per
default but can take the "test" and the "prod" values.

In test mode, some modules has to been neutralized for
example the payment acquirers, the delivery carrier, the
mail and so on..

During the loading of the modules, if the neuter-mode(#test)
is passed in command line, a method will search the neuter
function and apply the changes. In the case of the prod
mode, you want to revert everything you turn off. It's
why a "reverse_neuter" has to be defined too.

In this commit, I implement the neutering for the delivery
carrier. To do so, I add a new field "has_been_neutered" to
filter the records which must be reverted.