[IMP] base: Neutralize Mail Cron, IAP services
Add a neutralization routine for the mail, the cron
and the iap services.

Concerning the mail server and the cron, the same
routine as usual is used. There are archived and
the has_been_neutered field is added to these two
models, so they can be reactive.

Concerning the iap services, there are related to
'ir.config_parameters'. In this case, an
has_been_neutered field has no sense and there does
not exist a way to unactive the records. It's why
I add a "-neuter" flag to the key when I neutralize
them. Finally, the flag is removed when I come back
to production.