[IMP] website: improve compute speed of website_published
Without this patch, accessing a SO of ~2k lines through the website UI
(/my/orders/x) takes ~4.225s (of which 160ms postgres time).

With this patch, accessing the same SO takes ~3.272s (of which 153ms
postgres time) which means about 1s saved.

It is likely that any other business cases where `website_published` is
computed will also see a speed improvement.