[IMP] uom: flag international UoM
This commit introduces the `international_unit` allowing to flag unit of measure in odoo as UoM
recognized by the International System Of Units (https://en.wikipedia.org/wiki/International_System_of_Units).
All units are not introduced here, as we mostly handle UoM used in business flow (scientific ones are omitted
on purpose), but we included the most used 'derived' (cm for length, for instance). For a detailled list of the
internation derived units, see https://en.wikipedia.org/wiki/SI_derived_unit.

After a few search, the 'foot', 'ounce', ... are variable according to their countries and their measurement
system. They are not recognized by the SI.
Neither is the 'liter', but it correspond to a cubic decimetre, so we flag the odoo liter UoM as 'cubic decimeter'.

Once the field `international_unit` is added, we want to prevent the internation UoM deletion and modification:
- modification because, on the planet, 1 kilogram will be always be 1000 grams. Changing that might cause problem
during uom conversion business.
- deletion because we consider them as 'master data', since they are existing in the real world for all eternity
(or evidence to the contrary).

Task-1971491