[IMP] l10n_lu_reports: Individual Electrical Report Generation for Finance,Tax
Individual generation of xml file for each financial and tax reports for the
Luxembourg.
Before we are generating the report based on the wizard and combined
all the financial and tax report in single XML file. Instead we have provided individual
Export XML button in each Financial and Tax report for LU.