[REF] stock: should not create scrap if insufficient quantity
Before this commit, when a user creates scrap order and if the quantity is
insufficient, it will show insufficient wizard which will create scrap even if
we discard (cancel) it and increase the sequence and later on unlink the scrap
order.

After this commit, this behaviour is improved so that it will not create scrap
order in the first place.

To avoid creating scrap, added special button widget scrap_button, normal button
behaviour is to save record first and then call that button action but this
scrap_button will not save scrap record instead it directly calls inventory_data
method which will check if stock is sufficient or insufficient, if sufficient
stock then record of scrap is created and validated if insufficient stock then
inventory_data method will return insufficient wizard action, this button later
open that wizard if insufficient quantity.

task-1959647