[FIX] event: display timezone error
The issue is that, event widget on home page displays the event using a
calculation, so if the user is in a different time zone it shows the wrong event
date.For example a  event is on February 28 at 18:30 in Milwaukee but because it
is calculated, likely the Belgian server side data, it shows as March 1 because
in Belgium 28/2/2019 @ 18:30 CST is 1/3/2019 3:30 CET.
And also in-person event should never report a calculation on timezone since it
can be attended at event location and date-time should be according to that
event's timezone.

By this commit, this issue is fixed by using javascript approch.

task - 1944409