[FIX] mrp: Global timer for every users.
On the workorder views, all the productivity timer are displayed
for each users that works on the workorder. However the global
timer upper the details only shows the sum of productivity time
for the logged user. It would be a better information for manager
to have the sum of all the timer and have the total duration for the
workorder. Also the users still have their own timer in the details.

opw-2029830