[IMP] base: set booleans to False by default
This commit makes use of the ORM to set all boolean fields to `False` by
default, this implies that all boolean fields are implicitly
`required=True`, this of course is still overridable if desired.

This is to avoid things like domains containing bits like
['|', ('my_bool', '=', False), ('my_bool', '=', None)] and other
inconsistencies in methods like create that filter booleans with null
values and replace them with `False`.