[IMP] web_editor: review UI and drag and drop
This commit reviews the "DropBlock" related code to make it as clean
and functional as possible. The dropped snippets can now be dragged by
holding the whole overlay which shows up on mouseover (no more when
holding a button which showed up on click).

To do that part of the main UI had to be reviewed and some components
were then naturally improved:

- The code view is now resized automatically by design, no JS

- The popovers follow the content naturally, the same way as drag and
  drop UI. This was possible by making the *container* scroll, instead
  of the editable element.

- ...

Note: not everything is working because the low-level code does not
work (many fixme were introduced there), to check.