[ADD] web_editor: custom context menu with undo, redo, select all
Since not all features of the default context menu are available (eg.
"select all" doesn't trigger a specific event that we can then use to
trigger a select all in the Arch), we chose to do like Google Docs and
have our own custom context menu. For the sake of speed we didn't waste
time making it fully extensible yet but that wouldn't be a hard task to
accomplish.
Note: we couldn't include `paste` into the menu because 1) triggering a
`paste` event is impossible, for security reasons; 2) reading the
contents of the clipboard without a `paste` event requires the explicit
permission of the user. Google Docs gets around that on Chrome
because... well... it's their own browser and they trust themselves.
But their `paste` button on Firefox just opens a popup showing the
keyboard shortcut.