[FIX] web_editor: properly insert fragments with `Arch.insert`
This also fixes other bugs that the changes revealed, notably with the
removal of nodes (propagation to children was not working properly),
handling of spaces (now in `ArchNodeText`'s rules) and of `BR` tags
(now in `br`'s rules: add an extra `BR` if single BR at end of block).

Other minor change to the API as a result of these fixes:
`ArchNode.after` and `ArchNode.before` do not accept Arrays anymore but
take fragments (in line with the way `Arch.insert` works).