[IMP] web_editor: call plugin public methods as callback of `Arch.do`
When a plugin public method is called via a button or a hotkey, it is
now called as a callback of `Arch.do`. This means that those methods
are now all single transactions and that rendering will occur when they
are completed. This removes the need for calling `importUpdate` and
setting the range within those method. It also simplifies the API in
that it reduces the confusion between ArchNodes and ClonedArchNodes.
Further improvements should remove that confusion altogether.

Notes:
- Methods called as a callback of `Arch.do` can now return - besides
undefined to keep the default range behavior, or a range to set a
specific range - false to keep the range as it was before, or an Array
of ArchNodes (or their IDs) to set the range from the beginning of its
first item to the end of its last.
- In cases where we do not wish the method of a button to be called as
a callback of `Arch.do`, `PluginsManager.call` can be called with its
last argument set to `true`, or the button can have the argument
`no-transaction`. This can be useful for instance in the case of test
buttons and buttons that open modals.