[ADD] l10n_ar: Set short code of journal automatically for Argentinian Sale Journals
Since all the Argentinian sales journals use documents then the AFIP POS
Number is set but the short code of the journal is neglected: is not
used but in the same time is required for Odoo. For that reason we auto
fill the short name of the sales journal with the AFIP POS Number.