[IMP] web_editor: change default range in `Arch.do` to preserve previous
Callbacks of `Arch.do` that return nothing now preserve the range as it
was before the changes (or gives the illusion thereof). This should make
the development of higher level plugins easier as it abstracts the range
handling out of it.
Todo: harmonize that default behavior with the default outside of
`Arch.do`