[WIP] website: deleting attachments related to website to delete
When creating Website A and Website B, then customizing themes for both,
it creates attachements linked to both websites.
If we delete Website B, attachments linked to B are not deleted,
but their website_id is removed.
If the user then tries to customize theme for Website A, an error

raise ValueError("Expected singleton: %s" % self) ValueError:
Expected singleton: ir.attachment(...)

appears in the console, and it is impossible, for instance, to change colors.

opw 2035249
opw 2038414