fix sale_product_configurator_pricelist_tour
In a sale order, when adding a line or changing the price unit of a line,
onchange must update the amount_total

`onchange` dummy assign the values that changed:

```
        for name in todo:
            if name == 'id':
                continue
            value = record[name]
            field = self._fields[name]
            if field.type == 'many2one' and field.delegate and not value:
                # do not nullify all fields of parent record for new records
                continue
            record[name] = value
```

In the above case, only 'order_line' change on `sale.order`, nothing else,
which just assign `new` lines.
As we skip one2manys in _field_triggers, no changes was detected
when just changing 'order_line' in modified, and amount_total was therefore
not recomputed.

One way to overcome this would be to dummy assigned all field
of each one2many lines, but this will likely be unefficient.

Therefore I chose the easy way, another _field_triggers_onchange list,
which includes one2many specifically for this case.