[FIX] model.with_context() should not alter arguments in-place
Rev. fd2fb47028ff8491e3173491dd22b6ed0305102a tried to fix the
propagation of `allowed_company_ids` throughout `with_context()` calls.

That patch alters the arguments in-place as a side-effect, which can have
far-reaching consequences, e.g. when the given dict is resused for other
purposes or stored.
This was actually happening in `ir.filters.get_filters()`[1]: it passes
the orm_cache'd result of `res.users.context_get()` to `with_context()`,
resulting in a permanent alteration of the cached context. It caused
semi-random downstream errors.

This patch attempts to preserve the same logic and result without altering
the arguments as a side-effect.

[1] https://github.com/odoo/odoo/blob/b43e06dd5a80814e15ce20f50d683f0daaa19a93/odoo/addons/base/models/ir_filters.py#L60-L77