[FIX] account_facturx: support invoices received by email again
Revision 1a871d72438c731d08ef10e668b81b4ade900b85m refactored the code
to parse invoices received via attachments to support XML files, using
the attachment mimetype as discriminant.

However, messages received via message_post() can contain 2 kinds of
attachments: normal ir.attachment records, and namedtuples created by
the mail routing subsystem to hold the details of an attachment that is
not yet inserted in the database.

This last kind does not come with a mimetype attribute, and in any case
the code later tests the filename, so we can directly fallback to using
the filename as discriminant in those cases.

This avoids a crash when processing emails received from the mail
routing system, as the Attachment namedtuple type had not `mimetype`
attribute.