[REF] stock: stock.picking.type: multi company
First step of the multi-company revamp in stock: we would like the
company_id field of stock.picking to be a related of the company_id field
of its picking type. Not a related stored because our funky orm doesn't
support it (the ir rule check or constraints check is done before the
computed field are writtent in database) but we'll implement it another
way I suppose.

In order to do that, we need to make the company_id of the picking type
required. At the moment, it isn't required and is a related to the
company_id of its warehouse (which is required).

The thing is, warehouse_id is not set in case of dropshipping.

We now make company_id stored on the picking type, it's editable in the
view, proper domains are added on the field and consistency checks are
done at create and write.

We prevent editing the company_id once the record is created because
some pickings could be created etc, and it would be costly to implement
plus it's not really a use case (except create and edit directly but
we'll ignore that at the moment).

We also create create as many dropshipping picking types as there are
companies.

task-1985992