[REF] stock: res.partner: multi company
Check company consistency of property_stock_supplier and
property_stock_customer.

Add domains in the views.

Note that company_id of res.partner isn't required but the property
fields depends on the current user's company.

We also count on the fact the ir.rule will always be active and not via
an option in the general settings, a task that should be merged before
v13.

task-1985992