[REF] stock: stock.inventory: multi company
company_id on inventory line is now a classic m2o stored. It would have
been very nice to make it a stored related but it doesn't work with our
****** orm since stored related field are really "stored" after
validating the fields, which means all the constrains break. And if you
adapt your constrains as check in create/write, the access rule failed
miserably because company_id isn't stored yet, which means you have to
adapt your ir.rule to handle company_id falsy even if it is required on
the model (and then it doesn't crash but the contrains cannot be check
correctly).

task-1985992