[REF] stock: stock.production.lot: multi company
Add a required company_id on the stock.production.lot model.
Don't set a default so that we are sure that models creating
stock.production.lot explicitly pass the company. The drawback is that
creating a lot from the menu item is subpar.

TODO: change the company of a product after using it for a lot of
another company

task-1985992