[FIX] web: hide create button in search popup
purpose of this commit is to don't append a create button in the view dialog.

before this commit:
in mobile view, if we open searchcreate dialog, it opens its kanban view and
there we have create button (if not restricted from view arch) even if the user
has passed the option 'no_create' to true.

after this commit:
create button will be hidden if no_create is passed.

task- 49283