[FIX] auth_signup: prevent crash if http_routing not installed
Follow up of 269aa594111a152ad4b7714856ea745bfef57155

`request.lang` is only set on `http_routing` so it cannot be used here.

The fix here is the same as what was done on `web` on the mentioned commit.

To reproduce the issue:
- install any module, but not `http_routing`
- use reset password
- try to set a new password from the received link with token

A few notes about the mentioned commit:
- the `lang` `lazy_property` was removed because it was a string before and a
  record now, and lazy_property cannot return a record
- the `lang` on the `request` is now a record
- the `lang` on the `context` is still a string

closes odoo/odoo#36468

Signed-off-by: Romain Derie (rde) <rde@odoo.com>