[IMP] account, sale, purchase: Add multiple product lines in batch
We are providing options to add multiple product lines in batch with selecting
the multiple products in the search more options from the dropdown menu of
product.

We also provide multiple products add based on one Fancy button provided in
sales and purchase form view having kanban view of products. Clicking on any of
the product will add the quantity in the SO or PO lines. even you can reduce the
quantity also.