[FIX] account: make sure you can still change CoA
When you install accounting for a db with a specific
country on the company, the corresponding localization
will automatically be installed.  However, some
localizations have multiple CoAs, but they install
the first anyways with the idea you can still change
it afterwards.

Because of the closing entry for the tax report however
on installation of the CoA, an account move is created.

This made it not possible anymore to change the CoA as
there was already an account move.  Here, we decided to
make the constraint more supple and only avoid changing CoA
when the move / bank statement / payment has been confirmed.