[IMP] web: force master password
Force the user to set a master password before creating or restoring a
new database.
This will not prevent user from creating a database in command line or
via RPC but at least force the basic user to set a master password the
first time he access the database manager.