[FIX] payment: env shaenanigans during payment onboarding
The payment onboarding wizard can, depending on the values selected
in the wizard, install modules on the fly.

Before this commit, modules were installed one at a time and the
environment was not always reset at the right time between these
installations. This caused the values set in the wizard to not be
saved correctly on the target records (e.g. select Stripe in the
wizard, set values for the Stripe keys and submit the wizard: the
Stripe module got installed correctly but the values were not set
on the acquirer).

This commit regroups all module installations in a single operation
and use the new `with_env` method to reassign the environment of the
current wizard to avoid passing around environments and calling
methods on `self` with the wrong environment.