[IMP] account,*: Improve account.bank.statement.line / account.payment workflow
- Simplify the bank reconciliation.
- Simplify the management of "blue" lines in the reconciliation widget. A "blue" line is now
a journal item using a temporary liquidity account.
- Remove the bank reconciliation threshold date. The reconciliation report will show the not
already reconciled journal entries using a liquidity account and the not already reconciled
journal entries using a temporary liquidity account. Without accounting, an account.payment
will involve directly the liquidity account and then, will be considered as a statement line
directly.
- Remove the post_at bank reconciliation feature. The "paid" state will be set on the
invoices only if reconciled with a journal entry involving the journal's liquidity account.
With invoicing, the payment will do that so the "in_payment" state should never be shown up.
With accounting, only the statement lines have the power to move an invoice to the "paid"
state.
- Fix various corner cases about the management of multi-currency in bank statement lines.
- Fix the conversion dates in multi-currency: Since the bank/cash is always used on the
statement lines, it will use always the real "bank" date instead of the fictive payment one.
- Ensure the 'reconcile' method will raise an error if the involved moves are not posted.

--task: 2092096