[IMP] website_form: add new field below the selected one
This commit enables user to add a new field below selected one. For this, now
menu option to add new fields will be displayed on field rows instead of the
form itself, so that user can decide where he wants to add the new field.

This commit also copies design of selected field row and automatically applies
it on the newly added field to ease the designing.

task-1874350