[IMP] website_event_questions: improve question_ids from event.type
When applying an event.category on an event.event, we copy the event.questions from the event.type,
but we don't immediately copy the answers associated the these questions.

This was not done for technical reasons, we were limited by the JS framework.

Since this commit: 5e9f821808d697243d9aaaa7f6d4cce59bd40e58
We now correctly handle nested o2m when triggering onchange.

This commit cleans the code by correctly copying the answers along with the questions when changing
the event.type of an event.event.

PR #45497
Task 2189529