[REF] survey: remove question field, unnecessary related on title
PURPOSE

Purpose is to perform some cleaning in survey models: naming, ordering,
dead code. As new features are about to land in survey, notably live
interactions [1] and new survey building [2] performing a pre cleaning
is necessary.

SPECIFICATIONS

Question model holds two fields for its title, title and question. It
is mainly due to historical reasons, when updating models for certifications
and eLearning.

In this commit we keep only title field and remove the question related
field as it adds unnecessary complexity to the model.

LINKS

[0] Related to Task ID 2061901 (survey models cleaning and preparation)
[1] Task ID 1972640 (live interactions)
[2] Task ID 2119587 (new frontend for building surveys)

PR #40623