[MOV/REF] survey: split main controller into a "main" and a "survey"
This commit splits the main Survey controller into a "BaseSurveyController" and a "Survey"
controllers.

The "base" contains access methods as well as utility methods to compute the results displayed in
the statistics templates.

This is a preliminary work for the "survey live mode" task that will need to use those results in
a different template, and need to factorize the common methods.

Task#1972640