[FIX] stock: default date on stock move
When the user set the schedule date and then the moves. The default date
expected on the move is today. However two inverse function on picking
are trigger. _set_move_without_package and _set_scheduled_date.

In the case where _set_scheduled_date is trigger first it will not write
the date on the stock_move and the stock_move will be created afterward.
In this case the move will have today as date and the picking date will
remains in the future. That could be confusing for the user and
problematic for the forecast report that are focus on the move date.

In order to fix it, set a default_date on move that is equals to the
schedule date on its picking.

It remains an issue that _set_scheduled_date could trigger after
_set_move_without_package and would erase all the manual date from the
user. However it could not be fix without involve existing feature nor
adding a new button.

Task: 1849912