[IMP] survey : save questions on time limit reached
Purpose
=======

If the timer goes to 00:00, the form is automatically submitted but the answers
are not saved. The form should be submited, validation should filter out invalid
answers to avoid showing validation errors on the form, and save answers.
If the answer is not in the correct format : skip the question
If the answer is required but there is no answer : skip the question

This way, we can still compute score.

Specifications
==============

This commit removes the invalid answers from the question to save, only if the
timer is not over, to be able to compute score on not finished survey.

Task ID: 2152223
PR #41453