[REF] survey : use ajax instead redirection for screen transitions
This commit refactors the way the questions are loaded.
To go to next question(s), instead of redirecting to a new page containing
the next question(s), the submit process is loading the next question(s)
using ajax. The template of the question is prepared at server side and is
given in html string to the client that display the next question in a smooth
transition (fade in / fade out).

This commit includes also the start and the end screen of the survey inside
the survey form in order to perform smooth transition between start screen
and first question(s) and between last question(s) and end screen.

Only one master template (survey form) is now used to display every part of
the survey : start + questions + end screens.

The answer_done error is now handled by the _prepare_survey_data itself.
If the user wants to display an already_done survey, the returned template
is the end screen.

A new survey.user_input state has been introduced. This state is a transition
state to know where we are in the process of the survey. If the user arrives on
start screen, the state is 'New'. Once the user clicks on Start button, an rpc
call is made to survey/begin that set the state to 'Started' and returns the
survey_prepare_data that will display the first page after having set the state
to 'skip'.

Task ID: 2152223
PR #41453